Backend/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-allauth==65.5.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
google-auth==2.38.0
idna==3.10
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
urllib3==2.3.0
